+++ common/src/main/kotlin/app/shosetsu/common/consts/settings/SettingKey.kt
@@ -38,7 +38,7 @@ sealed class SettingKey<T : Any>(val name: String, val default: T) {
 	object ReaderTextSize : SettingKey<Float>("readerTextSize", 14f)
 
 	//TODO turn into float
-	object ReaderParagraphSpacing : SettingKey<Int>("readerParagraphSpacing", 1)
+	object ReaderParagraphSpacing : SettingKey<Float>("readerParagraphSpacing", 1f)
 	object ReaderIndentSize : SettingKey<Int>("readerIndentSize", 1)
 
 	/**
